Reduce length of App class using 2 files
